move vijf_slechte_kespen_naast_elkaar to Rakdeel
