+++ tests/test_rak.py
@@ -76,91 +76,3 @@ def test_kzg0202_gebreken(kzg0202_rak: Rak):
 #     ), f"Expected 1 non-numeric kesp hoogte, got {len(all_non_numeric)}: {all_non_numeric}"
 #     assert sum(all_hoogtes) == 1549, f"Expected total height of 1550 cm, got {sum(all_hoogtes)}"
 #     # expected = 1549 + 1"NM" (total len 97, numerical len 96)
-
-
-def test_vijf_slechte_kespen_naast_elkaar_met_mock(mock_rak_with_gebreken: Rak):
-    """Test vijf_slechte_kespen_naast_elkaar met mock data (K2-K6 zijn aangetast)."""
-
-    assert mock_rak_with_gebreken.vijf_slechte_kespen_naast_elkaar
-
-
-def test_vijf_slechte_kespen_naast_elkaar_minder_dan_vijf_kespen(mock_rak_with_gebreken: Rak):
-    """Test vijf_slechte_kespen_naast_elkaar met minder dan 5 kespen totaal."""
-
-    # Keep only K1-K3
-    mock_rak_with_gebreken.rakdelen[0].onderbouw.kespen = mock_rak_with_gebreken.rakdelen[0].onderbouw.kespen[:3]
-    assert not mock_rak_with_gebreken.vijf_slechte_kespen_naast_elkaar
-
-
-def test_vijf_slechte_kespen_naast_elkaar_onderbroken(mock_rak_with_gebreken: Rak):
-    """Test vijf_slechte_kespen_naast_elkaar met exact 4 consecutive aangetaste kespen."""
-
-    # K3 niet aangetast
-    mock_rak_with_gebreken.rakdelen[0].onderbouw.kespen[2].is_aangetast = False
-
-    assert not mock_rak_with_gebreken.vijf_slechte_kespen_naast_elkaar
-
-
-def test_vijf_slechte_kespen_naast_elkaar_gebrek(mock_rak_with_gebreken: Rak):
-    """Test vijf_slechte_kespen_naast_elkaar met een gebrek op K1 maar K1 zelf niet aangetast."""
-
-    # Maak K3 niet aangetast maar voeg een gebrek toe
-    mock_rak_with_gebreken.rakdelen[0].onderbouw.kespen[2].is_aangetast = False
-    mock_rak_with_gebreken.rakdelen[0].onderbouw.kespen[2].gebreken.append(
-        Gebrek(codering="GK1", omschrijving="Beschadigd kesp", figuurnummer="F2")
-    )
-
-    assert mock_rak_with_gebreken.vijf_slechte_kespen_naast_elkaar
-
-
-def test_vijf_slechte_kespen_naast_elkaar_meer_dan_vijf(mock_rak_with_gebreken: Rak):
-    """Test vijf_slechte_kespen_naast_elkaar met 6 consecutive aangetaste kespen."""
-
-    # K7 ook aangetast maken
-    mock_rak_with_gebreken.rakdelen[0].onderbouw.kespen[6].is_aangetast = True
-
-    assert mock_rak_with_gebreken.vijf_slechte_kespen_naast_elkaar
-
-
-def test_vijf_slechte_kespen_naast_elkaar_geen_kespen(mock_rak_with_gebreken: Rak):
-    """Test vijf_slechte_kespen_naast_elkaar met geen kespen."""
-    # Verwijder alle kespen
-    mock_rak_with_gebreken.rakdelen[0].onderbouw.kespen = []
-
-    assert not mock_rak_with_gebreken.vijf_slechte_kespen_naast_elkaar
-
-
-def test_vijf_slechte_kespen_naast_elkaar_next_rak(mock_rak_with_gebreken: Rak):
-    """Test vijf_slechte_kespen_naast_elkaar wanneer een volgende kesp in een nieuw rakdeel zit."""
-
-    # K1-K2 niet aangetast, K3-K7 aangetast
-    for i, kesp in enumerate(mock_rak_with_gebreken.rakdelen[0].onderbouw.kespen):
-        kesp.is_aangetast = i >= 1  # alleen de laatste 5 kespen worden aangetast
-        kesp.gebreken = []
-
-    # Check dat er nu 5 slechte kespen naast elkaar zijn
-    assert mock_rak_with_gebreken.vijf_slechte_kespen_naast_elkaar
-
-    # K3 ook niet aangetast maken
-    mock_rak_with_gebreken.rakdelen[0].onderbouw.kespen[2].is_aangetast = False
-
-    # Check dat er nu geen 5 slechte kespen naast elkaar zijn
-    assert not mock_rak_with_gebreken.vijf_slechte_kespen_naast_elkaar
-
-    # Voeg kesp toe, maar in volgende rakdeel
-    nieuw_rakdeel = mock_rak_with_gebreken.rakdelen[0].model_copy()
-    nieuw_rakdeel.rakdeel_id = "Constructie B"
-    nieuw_rakdeel.onderbouw.kespen = [
-        Kesp(
-            kesp_nummer="K8",
-            hoogte_cm=100,
-            breedte_cm=50,
-            lengte_uitstekend_deel_cm=30,
-            is_opsluitklos_aangetast=False,
-            is_aangetast=True,
-        )
-    ]
-    mock_rak_with_gebreken.rakdelen.append(nieuw_rakdeel)
-
-    # Check dat er nu nog steeds geen 5 slechte kespen naast elkaar zijn, omdat K8 in een nieuw rakdeel zit
-    assert not mock_rak_with_gebreken.vijf_slechte_kespen_naast_elkaar
